refactor: share extension package snapshot helpers
